R code - version 1
This is the first R code, so don't expect anything spectecular. It calculates the weekly averages (as it was asked in the assignmnet) and creates a new data frame that only contains these averages. Note that the comments are still in Hungarian (will be changed later). Also the code contains some unnecessary lines that helped me at the beginning (will delete those later). If you want to run the code yourself don't forget to change the working directory.